MLOps/MLOps_Innovative_Assignment_CAR_PRICE/requirements.txt
# ─── Core ML ─────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# ─── Experiment Tracking ─────────────────────
mlflow>=2.8.0

# ─── API (FastAPI + Uvicorn) ─────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ─── Authentication ──────────────────────────
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# ─── Database ────────────────────────────────
pymongo>=4.6.0

# ─── Monitoring ──────────────────────────────
prometheus-client>=0.20.0

# ─── Dashboard ───────────────────────────────
streamlit>=1.28.0
plotly>=5.18.0
requests>=2.31.0

# ─── Utilities ───────────────────────────────
python-docx>=0.8.11
